1. 修正GffHashGeneNCBI，将里面读取染色体的模块独立出来 2. SpeciesFile添加过滤染色体模块，如果一个物种含有太多染色体，如拼接序列，就会将太短的和Gff里面没有的染色体去除，这样方便后续MapSplice工作

+++ src/main/java/com/novelbio/analysis/seq/GeneExpTable.java
@@ -633,7 +633,7 @@ public class GeneExpTable {
 		} else if (enumExpression == EnumExpression.Counts) {
 			resultValue = value.intValue() + "";
 		} else if (enumExpression == EnumExpression.TPM) {
-			resultValue = (int)(value*mapreadsNum/allReadsNum) + "";
+			resultValue = value*mapreadsNum/allReadsNum + "";
 		} else if (enumExpression == EnumExpression.RPKM) {
 			resultValue = value*mapreadsNum*1000/allReadsNum/geneLen + "";
 		} else if (enumExpression == EnumExpression.UQRPKM) {
